refactor: add MFI endpoints to surveysApi.py
