+++ data_bridges_knots/client.py
@@ -13,19 +13,19 @@ import yaml
 from data_bridges_client.rest import ApiException
 from data_bridges_client.token import WfpApiToken
 
+from data_bridges_knots.endpoints.currencyApi import (
+    get_currency_list,
+    get_exchange_rates,
+    get_usd_indirect_quotation,
+)
+
 # from data.endpoints.prices import get_prices, get_exchange_rates
-from data_bridges_knots.endpoints.household import (
+from data_bridges_knots.endpoints.householdApi import (
     get_household_questionnaire,
     get_household_survey,
     get_household_surveys_list,
     get_household_xlsform_definition,
 )
-from data_bridges_knots.endpoints.currency import (
-    get_exchange_rates,
-    get_currency_list,
-    get_usd_indirect_quotation
-)
-
 from data_bridges_knots.helpers import get_adm0_code
 
 logname = "data_bridges_api_calls.log"
@@ -215,7 +215,6 @@ class DataBridgesKnots:
         logger.debug("Token used: %s", token.__repr__())
         return configuration
 
- 
     def get_commodities_list(
         self,
         country_iso3: Optional[str] = None,
@@ -394,7 +393,6 @@ class DataBridgesKnots:
                     f"Exception when calling CommodityUnitsApi->commodity_units_list_get: {e}"
                 )
                 raise
-
 
     def get_economic_indicator_list(
         self,
@@ -721,107 +719,6 @@ class DataBridgesKnots:
         choices["label"] = choices["choices"].apply(lambda x: x["label"])
         return choices[["name", "value", "label"]]
 
-    def get_mfi_surveys_full_data(
-        self, survey_id=None, page: Optional[int] = 1, page_size=20
-    ):
-        """
-        Get a full dataset that includes all the fields included in the survey in addition to the core Market Functionality Index (MFI) fields by Survey ID.
-        """
-        with data_bridges_client.ApiClient(
-            self._setup_configuration_and_authentication(self.config)
-        ) as api_client:
-            api_instance = data_bridges_client.SurveysApi(api_client)
-            env = self.env
-            try:
-                api_response = api_instance.m_fi_surveys_full_data_get(
-                    survey_id=survey_id,
-                    format="json",
-                    page=page,
-                    page_size=page_size,
-                    env=env,
-                )
-                logger.info("Successfully retrieved MFI surveys full data")
-                df = pd.DataFrame(api_response.items)
-                return df
-            except ApiException as e:
-                logger.error(
-                    f"Exception when calling SurveysApi->m_fi_surveys_full_data_get: {e}"
-                )
-                raise
-
-    def get_mfi_surveys(
-        self, adm0_code=0, page: Optional[int] = 1, start_date=None, end_date=None
-    ):
-        """
-        Retrieve Survey IDs, their corresponding XLS Form IDs, and Base XLS Form of all MFI surveys conducted in a country.
-        """
-        with data_bridges_client.ApiClient(
-            self._setup_configuration_and_authentication(self.config)
-        ) as api_client:
-            api_instance = data_bridges_client.SurveysApi(api_client)
-            env = self.env
-
-            try:
-                api_response = api_instance.m_fi_surveys_get(
-                    adm0_code=adm0_code,
-                    page=page,
-                    start_date=start_date,
-                    end_date=end_date,
-                    env=env,
-                )
-                logger.info("Successfully retrieved MFI surveys list")
-                df = pd.DataFrame([item.to_dict() for item in api_response.items])
-                df = df.replace({np.nan: None})
-                return df
-            except ApiException as e:
-                logger.error(
-                    f"Exception when calling SurveysApi->m_fi_surveys_get: {e}"
-                )
-                raise
-
-    def get_mfi_surveys_processed_data(
-        self,
-        survey_id=None,
-        page: Optional[int] = 1,
-        page_size=20,
-        format: Optional[str] = "json",
-        start_date=None,
-        end_date=None,
-        adm0_codes=None,
-        market_id=None,
-        survey_type=None,
-    ):
-        """
-        Get MFI processed data in long format.
-        """
-        with data_bridges_client.ApiClient(
-            self._setup_configuration_and_authentication(self.config)
-        ) as api_client:
-            api_instance = data_bridges_client.SurveysApi(api_client)
-            env = self.env
-
-            try:
-                api_response = api_instance.m_fi_surveys_processed_data_get(
-                    survey_id=survey_id,
-                    page=page,
-                    page_size=page_size,
-                    format=format,
-                    start_date=start_date,
-                    end_date=end_date,
-                    adm0_codes=adm0_codes,
-                    market_id=market_id,
-                    survey_type=survey_type,
-                    env=env,
-                )
-                logger.info("Successfully retrieved MFI surveys processed data")
-                df = pd.DataFrame([item.to_dict() for item in api_response.items])
-                df = df.replace({np.nan: None})
-                return df
-            except ApiException as e:
-                logger.error(
-                    f"Exception when calling SurveysApi->m_fi_surveys_processed_data_get: {e}"
-                )
-                raise
 
         # TODO: Get the scope and test these functions
         def get_rpme_base_data(
@@ -993,89 +890,6 @@ class DataBridgesKnots:
 
         # Add this function to the DataBridgesShapes class
 
-    def get_mfi_xls_forms(
-        self, adm0_code=0, page: Optional[int] = 1, start_date=None, end_date=None
-    ):
-        with data_bridges_client.ApiClient(
-            self._setup_configuration_and_authentication(self.config)
-        ) as api_client:
-            api_instance = data_bridges_client.XlsFormsApi(api_client)
-            env = self.env
-
-            try:
-                api_response = api_instance.m_fi_xls_forms_get(
-                    adm0_code=adm0_code,
-                    page=page,
-                    start_date=start_date,
-                    end_date=end_date,
-                    env=env,
-                )
-                logger.info("Successfully retrieved MFI XLS forms")
-                df = pd.DataFrame([item.to_dict() for item in api_response.items])
-                df = df.replace({np.nan: None})
-                return df
-            except ApiException as e:
-                logger.error(
-                    f"Exception when calling XlsFormsApi->m_fi_xls_forms_get: {e}"
-                )
-                raise
-
-    def get_mfi_xls_forms_detailed(
-        self, adm0_code=0, page: Optional[int] = 1, start_date=None, end_date=None
-    ):
-        """
-        Get a complete list of XLS Forms uploaded on the MFI Data Bridge in a given period of data collection.
-
-        Args:
-            adm0_code (int): Code for the country. Defaults to 0.
-            page (int): Page number for paged results. Defaults to 1.
-            start_date (str): Starting date for data collection range (YYYY-MM-DD format)
-            end_date (str): Ending date for data collection range (YYYY-MM-DD format)
-
-        Examples:
-            >>> client = DataBridgesShapes("data_bridges_api_config.yaml")
-            >>> # Get detailed XLS forms for country code 231
-            >>> detailed_forms = client.get_mfi_xls_forms_detailed(adm0_code=231)
-            >>> # Get forms within a date range
-            >>> forms_in_range = client.get_mfi_xls_forms_detailed(
-            ...     adm0_code=231,
-            ...     start_date="2023-01-01",
-            ...     end_date="2023-12-31"
-            ... )
-
-        Returns:
-            pandas.DataFrame: DataFrame containing XLS Forms data
-        """
-        with data_bridges_client.ApiClient(
-            self._setup_configuration_and_authentication(self.config)
-        ) as api_client:
-            api_instance = data_bridges_client.XlsFormsApi(api_client)
-            env = self.env
-
-            try:
-                api_response = api_instance.m_fi_xls_forms_get(
-                    adm0_code=adm0_code,
-                    page=page,
-                    start_date=start_date,
-                    end_date=end_date,
-                    env=env,
-                )
-                logger.info("Successfully retrieved detailed MFI XLS forms")
-
-                # Convert response items to DataFrame
-                df = pd.DataFrame([item.to_dict() for item in api_response.items])
-                df = df.replace({np.nan: None})
-
-                # Add total items count as DataFrame attribute
-                df.total_items = api_response.total_items
-
-                return df
-
-            except ApiException as e:
-                logger.error(
-                    f"Exception when calling XlsFormsApi->m_fi_xls_forms_get: {e}"
-                )
-                raise
 
     def get_prices(
         self,
@@ -1176,43 +990,7 @@ class DataBridgesKnots:
         df = df.replace({np.nan: None})
         return df
 
-    def get_mfi_surveys_base_data(
-        self, survey_id=None, page: Optional[int] = 1, page_size=20
-    ):
-        """
-        Get data that includes the core Market Functionality Index (MFI) fields only by Survey ID.
 
-        Args:
-            survey_id (int): Unique identifier for the collected data
-            page (int): Page number for paged results
-            page_size (int): Number of items per page
-
-        Examples:
-            >>> client = DataBridgesShapes("data_bridges_api_config.yaml")
-            >>> # Get MFI base data for a specific survey
-            >>> base_data = client.get_mfi_surveys_base_data(survey_id=123)
-
-        Returns:
-            pandas.DataFrame: DataFrame containing MFI base survey data
-        """
-        with data_bridges_client.ApiClient(
-            self._setup_configuration_and_authentication(self.config)
-        ) as api_client:
-            api_instance = data_bridges_client.SurveysApi(api_client)
-            env = self.env
-
-            try:
-                api_response = api_instance.m_fi_surveys_base_data_get(
-                    survey_id=survey_id, page=page, page_size=page_size, env=env
-                )
-                logger.info("Successfully retrieved MFI surveys base data")
-                return pd.DataFrame(api_response.items)
-
-            except ApiException as e:
-                logger.error(
-                    f"Exception when calling SurveysApi->m_fi_surveys_base_data_get: {e}"
-                )
-                raise
 
     def get_ipc_and_equivalent_data(self):
         pass
@@ -1234,8 +1012,8 @@ class DataBridgesShapes(DataBridgesKnots):
     def __init__(self, *args, **kwargs):
         warnings.warn(
             (
-                "\n[FUTURE WARNING]\n"
-                "DataBridgesShapes is deprecated and will be removed in v4.0.0.\n"
+                "\n[FUTURE WARNING]"
+                "DataBridgesShapes will be deprecated and be removed in v4.0.0 (July 2026).\n"
                 "Use 'DataBridgesKnots' instead.\n"
             ),
             FutureWarning,
@@ -1255,6 +1033,7 @@ class DataBridgesShapes(DataBridgesKnots):
             f"Brought to you with <3 by WFP VAM"
         )
 
+
 # Household Endpoints (IncubationApi)
 DataBridgesKnots.get_household_survey = get_household_survey
 DataBridgesKnots.get_household_surveys_list = get_household_surveys_list
@@ -1265,7 +1044,6 @@ DataBridgesKnots.get_household_questionnaire = get_household_questionnaire
 DataBridgesKnots.get_exchange_rate = get_exchange_rates
 DataBridgesKnots.get_currency_list = get_currency_list
 DataBridgesKnots.get_usd_indirect_quotation = get_usd_indirect_quotation
-
 
 
 if __name__ == "__main__":
