+++ data_bridges_knots/endpoints/currencyApi.py
@@ -22,9 +22,7 @@ logging.basicConfig(
 logger = logging.getLogger(__name__)
 
 
-def get_exchange_rates(
-    self, country_iso3: str, page_size: int = 1000
-) -> pd.DataFrame:
+def get_exchange_rates(self, country_iso3: str, page_size: int = 1000) -> pd.DataFrame:
     """Retrieves exchange rates for a given country from the Data Bridges API.
 
     Args:
@@ -62,14 +60,10 @@ def get_exchange_rates(
             env = self.env
 
             try:
-                api_exchange_rates = (
-                    api_instance.currency_usd_indirect_quotation_get(
-                        country_iso3=country_iso3, format="json", page=page, env=env
-                    )
+                api_exchange_rates = api_instance.currency_usd_indirect_quotation_get(
+                    country_iso3=country_iso3, format="json", page=page, env=env
                 )
-                responses.extend(
-                    item.to_dict() for item in api_exchange_rates.items
-                )
+                responses.extend(item.to_dict() for item in api_exchange_rates.items)
                 total_items = api_exchange_rates.total_items
                 logger.info("Fetching page %s", page)
                 max_item = page * page_size
@@ -137,10 +131,9 @@ def get_currency_list(
             return df
 
         except ApiException as e:
-            logger.error(
-                f"Exception when calling CurrencyApi->currency_list_get: {e}"
-            )
+            logger.error(f"Exception when calling CurrencyApi->currency_list_get: {e}")
             raise
+
 
 def get_usd_indirect_quotation(
     self,
